Moved Legend mixin into separate file.
