+++ js/modules/timeline.src.js
@@ -40,8 +40,7 @@ import H from '../parts/Globals.js';
 * @type {Highcharts.Series}
 */
 import Point from '../parts/Point.js';
-import L from '../parts/Legend.js';
-var LegendSymbolMixin = L.LegendSymbolMixin;
+import LegendSymbolMixin from '../mixins/legend-symbol.js';
 import U from '../parts/Utilities.js';
 var addEvent = U.addEvent, arrayMax = U.arrayMax, arrayMin = U.arrayMin, defined = U.defined, isNumber = U.isNumber, merge = U.merge, objectEach = U.objectEach, pick = U.pick, seriesType = U.seriesType;
 var TrackerMixin = H.TrackerMixin, Series = H.Series, seriesTypes = H.seriesTypes;
