+++ js/parts-map/ColorAxis.js
@@ -18,8 +18,8 @@ import H from '../parts/Globals.js';
 import Color from '../parts/Color.js';
 var color = Color.parse;
 import Point from '../parts/Point.js';
-import L from '../parts/Legend.js';
-var Legend = L.Legend, LegendSymbolMixin = L.LegendSymbolMixin;
+import Legend from '../parts/Legend.js';
+import LegendSymbolMixin from '../mixins/legend-symbol.js';
 import U from '../parts/Utilities.js';
 var addEvent = U.addEvent, erase = U.erase, extend = U.extend, isNumber = U.isNumber, merge = U.merge, pick = U.pick, splat = U.splat;
 import '../parts/Axis.js';
