+++ js/parts/Legend.js
@@ -1075,93 +1075,6 @@ var Legend = /** @class */ (function () {
     };
     return Legend;
 }());
-H.Legend = Legend;
-/**
- * Legend symbol mixin.
- *
- * @private
- * @mixin Highcharts.LegendSymbolMixin
- */
-H.LegendSymbolMixin = {
-    /**
-     * Get the series' symbol in the legend
-     *
-     * @private
-     * @function Highcharts.LegendSymbolMixin.drawRectangle
-     *
-     * @param {Highcharts.Legend} legend
-     *        The legend object
-     *
-     * @param {Highcharts.Point|Highcharts.Series} item
-     *        The series (this) or point
-     *
-     * @return {void}
-     */
-    drawRectangle: function (legend, item) {
-        var options = legend.options, symbolHeight = legend.symbolHeight, square = options.squareSymbol, symbolWidth = square ? symbolHeight : legend.symbolWidth;
-        item.legendSymbol = this.chart.renderer.rect(square ? (legend.symbolWidth - symbolHeight) / 2 : 0, legend.baseline - symbolHeight + 1, // #3988
-        symbolWidth, symbolHeight, pick(legend.options.symbolRadius, symbolHeight / 2))
-            .addClass('highcharts-point')
-            .attr({
-            zIndex: 3
-        }).add(item.legendGroup);
-    },
-    /**
-     * Get the series' symbol in the legend. This method should be overridable
-     * to create custom symbols through
-     * Highcharts.seriesTypes[type].prototype.drawLegendSymbols.
-     *
-     * @private
-     * @function Highcharts.LegendSymbolMixin.drawLineMarker
-     *
-     * @param {Highcharts.Legend} legend
-     *        The legend object.
-     *
-     * @return {void}
-     */
-    drawLineMarker: function (legend) {
-        var options = this.options, markerOptions = options.marker, radius, legendSymbol, symbolWidth = legend.symbolWidth, symbolHeight = legend.symbolHeight, generalRadius = symbolHeight / 2, renderer = this.chart.renderer, legendItemGroup = this.legendGroup, verticalCenter = legend.baseline -
-            Math.round(legend.fontMetrics.b * 0.3), attr = {};
-        // Draw the line
-        if (!this.chart.styledMode) {
-            attr = {
-                'stroke-width': options.lineWidth || 0
-            };
-            if (options.dashStyle) {
-                attr.dashstyle = options.dashStyle;
-            }
-        }
-        this.legendLine = renderer
-            .path([
-            'M',
-            0,
-            verticalCenter,
-            'L',
-            symbolWidth,
-            verticalCenter
-        ])
-            .addClass('highcharts-graph')
-            .attr(attr)
-            .add(legendItemGroup);
-        // Draw the marker
-        if (markerOptions && markerOptions.enabled !== false && symbolWidth) {
-            // Do not allow the marker to be larger than the symbolHeight
-            radius = Math.min(pick(markerOptions.radius, generalRadius), generalRadius);
-            // Restrict symbol markers size
-            if (this.symbol.indexOf('url') === 0) {
-                markerOptions = merge(markerOptions, {
-                    width: symbolHeight,
-                    height: symbolHeight
-                });
-                radius = 0;
-            }
-            this.legendSymbol = legendSymbol = renderer.symbol(this.symbol, (symbolWidth / 2) - radius, verticalCenter - radius, 2 * radius, 2 * radius, markerOptions)
-                .addClass('highcharts-point')
-                .add(legendItemGroup);
-            legendSymbol.isMarker = true;
-        }
-    }
-};
 // Workaround for #2030, horizontal legend items not displaying in IE11 Preview,
 // and for #2580, a similar drawing flaw in Firefox 26.
 // Explore if there's a general cause for this. The problem may be related
@@ -1169,7 +1082,7 @@ H.LegendSymbolMixin = {
 // elements.
 if (/Trident\/7\.0/.test(win.navigator && win.navigator.userAgent) ||
     isFirefox) {
-    wrap(H.Legend.prototype, 'positionItem', function (proceed, item) {
+    wrap(Legend.prototype, 'positionItem', function (proceed, item) {
         var legend = this, 
         // If chart destroyed in sync, this is undefined (#2030)
         runPositionItem = function () {
@@ -1185,8 +1098,5 @@ if (/Trident\/7\.0/.test(win.navigator && win.navigator.userAgent) ||
         }
     });
 }
-var exports = {
-    Legend: H.Legend,
-    LegendSymbolMixin: H.LegendSymbolMixin
-};
-export default exports;
+H.Legend = Legend;
+export default H.Legend;
