+++ ts/modules/solid-gauge.src.ts
@@ -89,10 +89,7 @@ declare global {
 
 import Color from '../parts/Color.js';
 const color = Color.parse;
-import L from '../parts/Legend.js';
-const {
-    LegendSymbolMixin
-} = L;
+import LegendSymbolMixin from '../mixins/legend-symbol.js';
 import U from '../parts/Utilities.js';
 const {
     clamp,
