+++ ts/modules/timeline.src.ts
@@ -144,10 +144,7 @@ declare global {
  */
 
 import Point from '../parts/Point.js';
-import L from '../parts/Legend.js';
-const {
-    LegendSymbolMixin
-} = L;
+import LegendSymbolMixin from '../mixins/legend-symbol.js';
 import U from '../parts/Utilities.js';
 const {
     addEvent,
