+++ ts/parts-more/PolygonSeries.ts
@@ -43,10 +43,7 @@ declare global {
     }
 }
 
-import L from '../parts/Legend.js';
-const {
-    LegendSymbolMixin
-} = L;
+import LegendSymbolMixin from '../mixins/legend-symbol.js';
 import U from '../parts/Utilities.js';
 const {
     seriesType
