+++ ts/parts/Legend.ts
@@ -18,35 +18,10 @@ import H from './Globals.js';
  */
 declare global {
     namespace Highcharts {
-        interface BubbleLegend extends LegendItemObject {
-        }
         interface LegendCheckBoxElement extends HTMLDOMElement {
             checked?: boolean;
             x: number;
             y: number;
-        }
-        interface LegendItemObject extends LegendSymbolMixin {
-            _legendItemPos?: Array<number>;
-            checkbox?: LegendCheckBoxElement;
-            checkboxOffset?: number;
-            itemHeight?: number;
-            itemWidth?: number;
-            legendGroup?: SVGElement;
-            legendItem?: SVGElement;
-            legendItems?: Array<SVGElement>;
-            legendItemHeight?: number;
-            legendItemWidth?: number;
-            legendLine?: SVGElement;
-            legendSymbol?: SVGElement;
-            pageIx?: number;
-        }
-        interface LegendSymbolMixin {
-            drawLineMarker(legend: Legend): void;
-            drawRectangle(legend: Legend, item: (Point|Series)): void;
-        }
-        interface Point extends LegendItemObject {
-        }
-        interface Series extends LegendItemObject {
         }
         interface SeriesOptions {
             legendType?: ('point'|'series');
@@ -131,7 +106,6 @@ declare global {
             public setText(item: (BubbleLegend|Point|Series)): void;
             public update(options: LegendOptions, redraw?: boolean): void;
         }
-        let LegendSymbolMixin: LegendSymbolMixin;
     }
 }
 
@@ -1668,141 +1642,6 @@ class Legend {
     }
 }
 
-H.Legend = Legend as any;
-
-/**
- * Legend symbol mixin.
- *
- * @private
- * @mixin Highcharts.LegendSymbolMixin
- */
-H.LegendSymbolMixin = {
-
-    /**
-     * Get the series' symbol in the legend
-     *
-     * @private
-     * @function Highcharts.LegendSymbolMixin.drawRectangle
-     *
-     * @param {Highcharts.Legend} legend
-     *        The legend object
-     *
-     * @param {Highcharts.Point|Highcharts.Series} item
-     *        The series (this) or point
-     *
-     * @return {void}
-     */
-    drawRectangle: function (
-        this: Highcharts.Series,
-        legend: Highcharts.Legend,
-        item: (Highcharts.Point|Highcharts.Series)
-    ): void {
-        var options = legend.options,
-            symbolHeight = legend.symbolHeight,
-            square = options.squareSymbol,
-            symbolWidth = square ? symbolHeight : legend.symbolWidth;
-
-        item.legendSymbol = this.chart.renderer.rect(
-            square ? (legend.symbolWidth - symbolHeight) / 2 : 0,
-            (legend.baseline as any) - symbolHeight + 1, // #3988
-            symbolWidth,
-            symbolHeight,
-            pick(legend.options.symbolRadius, symbolHeight / 2)
-        )
-            .addClass('highcharts-point')
-            .attr({
-                zIndex: 3
-            }).add(item.legendGroup);
-
-    },
-
-    /**
-     * Get the series' symbol in the legend. This method should be overridable
-     * to create custom symbols through
-     * Highcharts.seriesTypes[type].prototype.drawLegendSymbols.
-     *
-     * @private
-     * @function Highcharts.LegendSymbolMixin.drawLineMarker
-     *
-     * @param {Highcharts.Legend} legend
-     *        The legend object.
-     *
-     * @return {void}
-     */
-    drawLineMarker: function (
-        this: Highcharts.Series,
-        legend: Highcharts.Legend
-    ): void {
-
-        var options = this.options,
-            markerOptions = options.marker,
-            radius,
-            legendSymbol,
-            symbolWidth = legend.symbolWidth,
-            symbolHeight = legend.symbolHeight,
-            generalRadius = symbolHeight / 2,
-            renderer = this.chart.renderer,
-            legendItemGroup = this.legendGroup,
-            verticalCenter = (legend.baseline as any) -
-                Math.round((legend.fontMetrics as any).b * 0.3),
-            attr = {} as Highcharts.SVGAttributes;
-
-        // Draw the line
-        if (!this.chart.styledMode) {
-            attr = {
-                'stroke-width': options.lineWidth || 0
-            };
-            if (options.dashStyle) {
-                attr.dashstyle = options.dashStyle;
-            }
-        }
-
-        this.legendLine = renderer
-            .path([
-                'M',
-                0,
-                verticalCenter,
-                'L',
-                symbolWidth,
-                verticalCenter
-            ])
-            .addClass('highcharts-graph')
-            .attr(attr)
-            .add(legendItemGroup);
-
-        // Draw the marker
-        if (markerOptions && markerOptions.enabled !== false && symbolWidth) {
-
-            // Do not allow the marker to be larger than the symbolHeight
-            radius = Math.min(
-                pick(markerOptions.radius, generalRadius),
-                generalRadius
-            );
-
-            // Restrict symbol markers size
-            if ((this.symbol as any).indexOf('url') === 0) {
-                markerOptions = merge(markerOptions, {
-                    width: symbolHeight,
-                    height: symbolHeight
-                });
-                radius = 0;
-            }
-
-            this.legendSymbol = legendSymbol = renderer.symbol(
-                this.symbol as any,
-                (symbolWidth / 2) - radius,
-                verticalCenter - radius,
-                2 * radius,
-                2 * radius,
-                markerOptions
-            )
-                .addClass('highcharts-point')
-                .add(legendItemGroup);
-            legendSymbol.isMarker = true;
-        }
-    }
-};
-
 // Workaround for #2030, horizontal legend items not displaying in IE11 Preview,
 // and for #2580, a similar drawing flaw in Firefox 26.
 // Explore if there's a general cause for this. The problem may be related
@@ -1812,7 +1651,7 @@ if (
     /Trident\/7\.0/.test(win.navigator && win.navigator.userAgent) ||
     isFirefox
 ) {
-    wrap(H.Legend.prototype, 'positionItem', function (
+    wrap(Legend.prototype, 'positionItem', function (
         this: Highcharts.Legend,
         proceed: Function,
         item: (Highcharts.Point|Highcharts.Series)
@@ -1835,9 +1674,6 @@ if (
     });
 }
 
-const exports = {
-    Legend: H.Legend,
-    LegendSymbolMixin: H.LegendSymbolMixin
-};
+H.Legend = Legend as any;
 
-export default exports;
+export default H.Legend;
