+++ ts/parts/Series.ts
@@ -690,10 +690,7 @@ declare global {
 
 ''; // detach doclets above
 
-import L from './Legend.js';
-const {
-    LegendSymbolMixin
-} = L;
+import LegendSymbolMixin from '../mixins/legend-symbol.js';
 import Point from './Point.js';
 import U from './Utilities.js';
 const {
